+++ packages/config-updater/src/index.ts
@@ -1,2 +1,6 @@
 export { ConfigUpdater } from "./ConfigUpdater.js";
+export {
+	ensureGitHubCredentialHelper,
+	handleGitHubTokens,
+} from "./handlers/githubTokens.js";
 export * from "./types.js";